+++ board.rb
@@ -1,3 +1,34 @@
+require './ship'
 class Board
+  def initialize
+    @ship
+    @placed
+    @fleet = []
+  end
+
+  def has_ship_on?(x_axis, y_axis)
+    ship_on = false
+    @fleet.each do |ship|
+      if ship.covers?(x_axis, y_axis)
+        ship_on = true
+      end
+    end
+      return ship_on
+  end
+  def place_ship(ship, x_axis, y_axis, across)
+    ship.place(x_axis, y_axis, across)
+    conflict = false
+    @fleet.each do |ship|
+      if ship.overlaps?(ship)
+        conflict = true
+      end
+    end
+    if !conflict
+      @fleet << ship
+      return true
+    else
+      return false
+    end
+  end
 
 end
